requirements.txt
Django==2.0a1
django-debug-toolbar==1.8
django-model-utils==3.0.0
ipython==6.2.0